test(core): split multi-describe token-graph + load-layered test files
Enforce the one-describe-per-file convention on three core test files
that had drifted from it:

- build.test.ts -> build-{extract-writes,baseline,affected-by,diagnostics}.test.ts
- walk.test.ts -> walk-resolve.test.ts (leaf resolver, incl. the stray
  top-level it + edge cases) and walk-resolve-alias.test.ts
  (alias-preserving variants)
- load-layered.test.ts -> load-layered.test.ts +
  load-layered-disabled-axes.test.ts + load-layered-alias-switch.test.ts,
  with shared fixture config lifted into _load-layered-helpers.ts

Justified beforeAll perf escape-hatches (~1s loadProject) preserved. No
test behavior change; 220 core tests still green.
